code/requirements.txt
numpy==2.2.4
torch==2.6.0
torchvision==0.21.0
pyyaml==6.0.2